fix(ui): hide providers without active mappings
